requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.5.2 